Move browser.js into squiffy-editor.js
