+++ index.html
@@ -17,7 +17,6 @@
     <script src="lib/bootbox.js"></script>
     <script src="lib/chosen.jquery.min.js"></script>
     <script type="module" src="/src/squiffy-editor.js"></script>
-    <script type="module" src="/src/browser.js"></script>
   </head>
   <body>
     <div id="squiffy-editor">
